+++ src/ResourceManager/RecoveryServices.SiteRecovery/Commands.RecoveryServices.SiteRecovery/Policy/GetAzureRmRecoveryServicesAsrPolicy.cs
@@ -30,7 +30,7 @@ namespace Microsoft.Azure.Commands.RecoveryServices.SiteRecovery
         "AzureRmRecoveryServicesAsrPolicy",
         DefaultParameterSetName = ASRParameterSets.Default)]
     [Alias("Get-ASRPolicy")]
-    [OutputType(typeof(IEnumerable<ASRPolicy>))]
+    [OutputType(typeof(ASRPolicy))]
     public class GetAzureRmRecoveryServicesAsrPolicy : SiteRecoveryCmdletBase
     {
         /// <summary>