+++ src/ResourceManager/RecoveryServices.SiteRecovery/Commands.RecoveryServices.SiteRecovery/RecoveryPoint/GetAzureRmRecoveryServicesAsrRecoveryPoint.cs
@@ -30,7 +30,7 @@ namespace Microsoft.Azure.Commands.RecoveryServices.SiteRecovery
         "AzureRmRecoveryServicesAsrRecoveryPoint",
         DefaultParameterSetName = ASRParameterSets.ByObject)]
     [Alias("Get-ASRRecoveryPoint")]
-    [OutputType(typeof(IEnumerable<ASRRecoveryPoint>))]
+    [OutputType(typeof(ASRRecoveryPoint))]
     public class GetAzureRmRecoveryServicesAsrRecoveryPoint : SiteRecoveryCmdletBase
     {
         /// <summary>